requirements.txt
pycryptodome
pyotp
